vision-research-server/pyproject.toml
[project]
name = "vision-research-server"
version = "0.1.0"
description = "Vision Research MCP Server — YOLO detection + Claude Vision enrichment"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.127.1",
    "fastmcp>=2.14.1",
    "uvicorn>=0.40.0",
    "anthropic>=0.75.0",
    "ultralytics>=8.3.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
]
